requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
sentence-transformers
pypdf
numpy
pytest
openai